requirements.txt
langgraph>=0.2.0
langchain-core>=1.2.8
langchain-groq>=1.1.0
groq>=0.37.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
